+++ test/RouteTest.php
@@ -13,7 +13,7 @@ use Zend\Expressive\Router\Exception\InvalidArgumentException;
 use Zend\Expressive\Router\Route;
 
 /**
- * @covers Zend\Expressive\Router\Route
+ * @covers \Zend\Expressive\Router\Route
  */
 class RouteTest extends TestCase
 {
@@ -166,6 +166,8 @@ class RouteTest extends TestCase
 
     /**
      * @dataProvider invalidHttpMethodsProvider
+     *
+     * @param array $invalidHttpMethods
      */
     public function testThrowsExceptionIfInvalidHttpMethodsAreProvided(array $invalidHttpMethods)
     {